lib: provide fdelay_add_ps for fdelay_config_pulse_ps
In fdelay_config_pulse_ps, the pulse length is given in
picoseconds but, for some reason, computation of the
pulse end time by addition of that length was commented
out.

This requires fdelay_add_ps to be implemented. There is a
similar function in fine-delay.h, but it is only accessible
to kernel-space code and is particular to zio attributes.