zio: new upstream commit
This brings in a few fixes in the "for-linux-2.6.32" branch.  In
particular, this solves the issue with too-fast writes to the output
channels using the "user" trigger. Later fine-delay commits take
advantage of it.  Also, this allows the input csets to mark a hardware
timestamp in the trigger, and have it preserved in the final control
block.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>