lib/fdelay-init: use another naming pattern
There are 4 possible combinations for device names.

They can be in /dev or in /dev/zio according to which udev version
you run: older ones didn't create the directory.

Moreover the device name can be called "zio-fd" or just "fd" according
to which version of ZIO you run: when we backported to 2.6.24 we dropped
the leading "zio-" part because old kernels had a short limit on device
names. Unfortunatelty fine-delay is being used on 2.6.24 as well.