sdb-lib: copied from fpga-config-space
The SDB library is an exact copy of the sdbfs/lib directory of repo:
   git://ohwr.org/hdl-core-lib/fpga-config-space.git
at version "sdb-v1.1" (commit de2c98f).

It actually is the same code that is already in fmc-bus/sdb-lib, but
until I finalize eeprom SDB support within fmc-bus I need a copy here.
This directory will disappear when fmc-bus is completed in that respect,
but we need a release with SDB in the eeprom now for fine-delay.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>