fd-core: ugly hack to correct ugly base addr hack
This ugly hack (hardcoding the base address of slot 1 fdelay
core) is needed in the SVEC, as we have yet no means to discover
that core's base address via SDB (and will never have, probably).
The slot number of the core would probably better be taken from
a register in the core itself.