lib: fdelay_open: don't set post-samples if not needed
Changing critical trigger parameters forces an abort of pending
blocks.  This means that a running reader will loos its own active
block when lib/fdelay-pulse is called (the incoming samples are not
lost, but the reader must be restarted).

The previous code would work with the new self-timed feature in recent
zio master, but by now this patch allows remaining at zio-1.0 -- in
the "fixes" branch.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>