tasklet: proceed with data_done even if no block is there
Now that we are self-timed, the only reason white the active input
block is null may be a zio buffer-full situation.  We need to keep
the loop running anyways (by discarding one sample), because we miss
notification when the buffer has new space.

We can't change zio to this aim, as it would be extra code in a hot
path: we'd better loose one block when the buffer is full, we would
loose it soon after anyways, since the system is overloaded.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>